requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
gtts>=2.5
rapidfuzz>=3.6
requests>=2.31
beautifulsoup4>=4.12
